[DFT] DFT Regeneration and test for mcp_pmsm_foc_encoder_pic32_mk_mclv2_pic32_mk_mca_pim
MH3-71442

+++ apps/mcp_pmsm_foc_encoder_pic32_mk/firmware/src/config/mclv2_pic32_mk_mca_pim/peripheral/qei/plib_qei1.c
@@ -91,13 +91,13 @@ void QEI1_Initialize (void)
 void QEI1_Start(void)
 {
     /* Enable QEI channel */
-    QEI1CON |= (uint32_t)_QEI1CON_QEIEN_MASK;
+    QEI1CON |= _QEI1CON_QEIEN_MASK;
 }
 
 void QEI1_Stop(void)
 {
     /* Disable QEI channel */
-    QEI1CON &= ~(uint32_t)_QEI1CON_QEIEN_MASK;
+    QEI1CON &= ~_QEI1CON_QEIEN_MASK;
 }
 
 uint32_t QEI1_PulseIntervalGet(void)
